split mongo tests
